Fix for test Query to check if connection is good patch.
Notice that the patch was missing the ! in front. Must have hit a space and erased the ! in front of the statement by mistake before committing code.

Description of original patch below with some changes:

Modification to send a test query to check if the connection is good before execution of all database query. 

Patch for databases that can't be setup to use a persistent connection or where the network is not always connected(i.e. mobile 3G networks).

Notes:

1. Do not add this code if you are using Derby since it is a local database and is not needed. It will only add a delay of 3ms to each query.
2. Delay time introduced by the additional test query is close to that of the ping times to the database server.

For example if it takes 10 queries to complete a user request then:

if ping time = 20ms then delay time of 1/5 second is added to the query time.
if ping time = 100ms then delay time of 1 second is added to the query time.
if ping time = 500ms then delay time of 5 seconds is added to the query time.